Extract Emails module into new component.
